Moved MXRUS code into a separate dedicated directory
MXRUS code has nothing to do with the rest of the SDK, it's primarily a UGC creation and loading module. It now has a dedicated directory that doesn't reference the rest of the SDK.

Part of the motivation behind this separation is the possibility that we might want to make MXRUS into a separate repository altogether in the future.